fix: hook useMediaQuery moved to a new folder
